Added client implementation in Python.
The functionality is backwards compatible with the client written in Bash.

New functionality:

Getting State
A curl command can be used to obtain a status value of OK/WARNING/CRITICAL based on parameters in the URL's query. For example:

    # curl -X GET  http://iperf-client-router/state/server=iperf-server-service,port=5001,warning=4000m,critical=3000m

In the above case, if the iperf results are lower than 3,000 M-bit/sec, then the command will return "CRITICAL". If the iperf results are faster than 3,000 M-bit/sec but slower than 4,000 M-bit/sec, then the command will return "WARNING". If the iperf results are faster than 4,000 M-bit/sec then the command will return "OK".

The following binary prefixes can be used:

- k (K-bit)
- m (M-bit)
- g (G-bit)
- t (T-bit)